wip: create an API model and use that for codegen
That is, instead of operating directly on the OpenAPI struct,
first transform that into a dedicated flattened model, and
then operate on that model.

The downside is that it's a fairly substantial rewrite, and I'm
still only partway through implementing it. Note that everything
related to actual codegen is commented out fornow.

The upside is that by operating on a flattened model, it turns
out to be _much_ easier to accomplish goals like infallible
refenence-indexing, ident deduplication, and so forth. We've
already eliminated the function which was marked as "weird and bad";
it's already been replaced with code which is simpler and more resilient.

More to do here, but I'm happy with the direction this is taking.