separate treatment of specification name and rust name
It turns out that we can't just use a single 'name' and always do
the right thing with it; we need to keep separate track of the
name as defined in the specification, and the Rust type name which
we've derived for an item.

However, once we've done that, we can get some nice properties, such
as deconflicting with types in the Rust prelude. This is also a
prerequisite for features like automatic renaming.