propagate well-known types into the system
Previously, we'd been operating purely on the schema level. This caused
problems when schemas couldn't be resolved for one reason or another.

Now, we discover well-known types and propagate them while constructing
the `ApiModel`. In addition, we no longer omit defined but unresolved
schemas. Instead, we use an arbitrary `serde_json::Value` for those types.