wip: generate structured request objects
Requests need to be generated as public items. This means that we
need the ability to generate items which fundamentally are public
typedefs to other generated items.

Once we have that ability, doing the right thing for requests breaks
down to doing the right thing at varying levels of reference, plus
a bit of special handling to ensure that the generated name (which
is part of the public interface) is reasonably human-friendly, and
the request type is as simple as possible while handling all necessary
cases.