fix resolve trait implementation for borrowed values
Turns out that there's no good way to directly implement Resolve
for both an owned and a borrowed reference by defining one in terms
of the other. That's ok; the solution is to define both of those in
terms of a helper function.

Just needed to rework the macro to actually produce that helper function,
which implied embedding the output in a small private module... but now
everything Just Works.