pass around additional parse parameters so we can detect `allOf` singletons
Appropriately detecting whether a schema meets all the criteria of
an `allOf` singleton requires at least some knowledge of the outer
scope.

However, we call `is_property_singleton` deep within a complex mutually
recursive call graph. This means that we need to pass around all its
required parameters pervasively throughout that call graph.

So, we do that.

This commit doesn't actually do anything with the singleton; it just
fixes the calls such that it is possible to do the right thing there.