fix the problem
The problem was an early return in the response generation code,
which occurred because there was a bit of model confusion going on.
We were modeling partial responses as `Item`s, which turns out to
be the wrong abstraction. We need to sometimes use the response
in its entirety, but we also need to incorporate it into a larger
enum sometimes.

The fix came in two parts:

- create a new model of a response item in the `ApiModel`
- rewrite the response-handling code to use that new item

At this point, it looks like we've successfully fixed the issue--
the code generator is emitting what looks like correct output given
the input case--but I've somehow borked the test case generator, so
that isn't running any cases anymore. That needs fixing.